delete customer logic fixed

+++ internal/core/usecases/remove_customer.go
@@ -1,6 +1,7 @@
 package usecases
 
 import (
+	"fmt"
 	"log"
 
 	"github.com/emersonnobre/tica-api-go/internal/core/repositories"
@@ -18,6 +19,12 @@ func NewRemoveCustomerUseCase(repository repositories.CustomerRepository) *Remov
 }
 
 func (u *RemoveCustomerUseCase) Execute(id int) types.UseCaseResponse {
+	count, _ := u.repository.GetCount(fmt.Sprintf("WHERE active = TRUE AND id = %d", id))
+	if count == 0 {
+		message, errorName := "Cliente não encontrado!", types.GetNotFoundErrorName()
+		return types.NewUseCaseResponse(nil, &errorName, &message)
+	}
+
 	if err := u.repository.Delete(id); err != nil {
 		log.Println("ERROR:", err)
 		message, errorName := "Erro ao remover cliente!", types.GetInternalErrorName()
